[search-params] testing
