+++ server/test/messages.template.test.ts
@@ -257,4 +257,44 @@ describe('Templates: Messages', () => {
         );
       });
   });
+
+  test('Datatypes template', () => {
+    return request(server)
+      .get('/api/messages/template/datatypes')
+      .expect(200)
+      .expect((resp) => {
+        const compiled = _.template(resp.body);
+
+        expect(
+          compiled({
+            name: 'widget',
+            version: '0.0.2',
+            schema: {
+              $id: 'https://example.com/widget.schema.json',
+              $schema: 'https://json-schema.org/draft/2020-12/schema',
+              title: 'Widget',
+              type: 'object',
+              properties: {
+                id: {
+                  type: 'string',
+                  description: 'The unique identifier for the widget.',
+                },
+                name: {
+                  type: 'string',
+                  description: "The person's last name.",
+                },
+              },
+            },
+          }),
+        ).toBe(
+          formatTemplate(`
+            const datatype = await firefly.createDatatype({
+              name: 'widget',
+              version: '0.0.2',
+            }, {"$id":"ht ...  name."}}}) ;
+            return { type: 'datatype', id: datatype.id };
+        `),
+        );
+      });
+  });
 });
